requirements.txt
django-classy-tags
